requirements.txt
certifi==2018.4.16
chardet==3.0.4
cloudpickle==0.5.3
cycler==0.10.0
dask==0.17.5
decorator==4.3.0
h5py==2.9.0
idna==2.6
joblib==0.13.2
kiwisolver==1.0.1
matplotlib==2.2.2
networkx==2.1
numpy==1.14.3
opencv-python==3.4.1.15
pbr==4.0.3
Pillow==5.1.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
requests==2.18.4
scikit-image==0.14.0
scikit-learn==0.21.2
scipy==1.1.0
six==1.11.0
stevedore==1.28.0
toolz==0.9.0
torch==1.1.0
torchfile==0.1.0
torchvision==0.3.0
tornado==5.0.2
tqdm==4.23.4
typing==3.6.4
urllib3==1.22
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
visdom==0.1.8.3
websocket-client==0.47.0
